AGENT-56 Add subsystem tests for agent-based installer
Added subsystem tests to read in ZTP manifests to create cluster
and infraenv.

Moved documentation about the agent-based installer and how
the client is used to docs/agent-based-installer.md.

Renamed ephemeral installer to agent-based installer to align
with recent naming decision.
